Migrate middlwares to programmatic for plugin that do not need to export them
