+++ examples/getstarted/package.json
@@ -14,6 +14,7 @@
   "dependencies": {
     "@strapi/admin": "3.6.8",
     "@strapi/plugin-documentation": "3.6.8",
+    "@strapi/plugin-sentry": "3.6.8",
     "@strapi/plugin-graphql": "3.6.8",
     "@strapi/plugin-i18n": "3.6.8",
     "@strapi/plugin-users-permissions": "3.6.8",
