+++ packages/core/upload/server/register.js
@@ -1,9 +1,13 @@
 'use strict';
 
-const registerUploadMiddlware = require('./controllers/upload-middleware');
+const registerUploadMiddlware = require('./middlewares/upload');
 
+/**
+ * Register upload plugin
+ * @param {{ strapi: import('@strapi/strapi').Strapi }}
+ */
 module.exports = async ({ strapi }) => {
-  await registerUploadMiddlware();
+  await registerUploadMiddlware({ strapi });
 
   if (strapi.plugin('graphql')) {
     require('./graphql')({ strapi });
