+++ packages/plugins/sentry/server/index.js
@@ -1,0 +1,13 @@
+'use strict';
+
+const register = require('./register');
+const bootstrap = require('./bootstrap');
+const services = require('./services');
+const config = require('./config');
+
+module.exports = () => ({
+  register,
+  bootstrap,
+  config,
+  services,
+});
