--- packages/plugins/sentry/server/middlewares/index.js
@@ -1,7 +1,0 @@
-'use strict';
-
-const sentry = require('./sentry');
-
-module.exports = {
-  sentry,
-};
