+++ packages/plugins/sentry/strapi-server.js
@@ -1,13 +1,3 @@
 'use strict';
 
-const bootstrap = require('./server/bootstrap');
-const services = require('./server/services');
-const middlewares = require('./server/middlewares');
-const config = require('./server/config');
-
-module.exports = () => ({
-  bootstrap,
-  config,
-  middlewares,
-  services,
-});
+module.exports = require('./server');
